requirements.txt
aiogram==2.25.1
requests
psycopg2-binary
